+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added console features for camera devices
+
 ## [3.3.2] - 2018-11-08
 ### Changed
  - Fail gracefully when loading a device and the port is not available